+++ docs/CHANGELOG.md
@@ -2,6 +2,7 @@
 * Fixes password problems when signing in.
 * Prevents user from registering with invalid passwords.
 * Disables iPad sleep timer.
+* Moves from Balanced to Stripe for credit card processing.
 
 ## 3.4.0 April 13 2015
 * Fixes problems entering in phone numbers.